+++ CHANGELOG.md
@@ -1,3 +1,18 @@
+### Additions
+* Added quick buttons for navigating the EMI config
+* Added configurable subpanels for sidebars
+* Added new sidebar types
+    * Craft History - A live updating history of the crafts you've performed
+	* Lookup History - A live updating history of the stacks you've looked at recipes for
+* Added full support for chance in recipes and the recipe tree
+
+### Tweaks
+* Craftables now display their output count, to help disambiguate them
+
 ### Fixes
-* Fixed handling of null tool materials in vanilla plugin
-* Added more safety to late recipe registering
+* Sidebar background should properly draw behind GUIs with overlaps
+* Chance support, #104 #105
+
+### API
+* Added convenience method for easier text widget construction
+* Added sorting method for recipes based on identifier